Trickle Ice with passing candidates, structured close reasons, refactor tunneling configuration: update imports, remove unused parameters, and enhance cleanup logic
