+++ keepercommander/commands/discoveryrotation.py
@@ -9,26 +9,18 @@
 # Contact: sm@keepersecurity.com
 #
 import argparse
-import asyncio
 import fnmatch
 import json
 import logging
 import os.path
 import re
-import queue
-import sys
-import threading
 import time
 from datetime import datetime
 from typing import Dict, Optional, Any, Set, List
 
 
 import requests
-from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives import hashes
-from cryptography.hazmat.primitives.ciphers.aead import AESGCM
-from cryptography.hazmat.primitives.kdf.hkdf import HKDF
-from keeper_secrets_manager_core.utils import url_safe_str_to_bytes, bytes_to_base64, base64_to_bytes
+from keeper_secrets_manager_core.utils import url_safe_str_to_bytes
 
 from .base import (Command, GroupCommand, user_choice, dump_report_data, report_output_parser, field_to_title,
                    FolderMixin, RecordMixin, register_pam_legacy_commands)
@@ -52,16 +44,14 @@ from .pam.router_helper import router_send_action_to_gateway, print_router_respo
     router_get_connected_gateways, router_set_record_rotation_information, router_get_rotation_schedules, \
     get_router_url
 from .record_edit import RecordEditMixin
-from .tunnel.port_forward.endpoint import WebRTCConnection, TunnelEntrance, READ_TIMEOUT, \
-    find_open_port, CloseConnectionReasons, SOCKS5Server, TunnelDAG, get_config_uid, MAIN_NONCE_LENGTH, \
-    SYMMETRIC_KEY_LENGTH, get_keeper_tokens
+from .tunnel.port_forward.TunnelGraph import TunnelDAG
+from .tunnel.port_forward.tunnel_helpers import get_config_uid, get_keeper_tokens
 from .. import api, utils, vault_extensions, crypto, vault, record_management, attachment, record_facades
 from ..display import bcolors
 from ..error import CommandError, KeeperApiError
 from ..params import KeeperParams, LAST_RECORD_UID
 from ..proto import pam_pb2, router_pb2, record_pb2
-from ..subfolder import find_folders, find_parent_top_folder, \
-    try_resolve_path, BaseFolderNode
+from ..subfolder import find_parent_top_folder, try_resolve_path, BaseFolderNode
 from ..vault import TypedField
 from ..discovery_common.record_link import RecordLink
 from .discover.job_start import PAMGatewayActionDiscoverJobStartCommand
@@ -87,6 +77,7 @@ from .pam_saas.user import PAMActionSaasUserCommand
 from .pam_saas.remove import PAMActionSaasRemoveCommand
 from .pam_saas.config import PAMActionSaasConfigCommand
 from .pam_saas.update import PAMActionSaasUpdateCommand
+from .tunnel_and_connections import PAMTunnelCommand, PAMConnectionCommand, PAMSplitCommand
 
 
 # These characters are based on the Vault
@@ -126,28 +117,6 @@ class PAMGatewayCommand(GroupCommand):
         # self.register_command('connect', PAMConnect(), 'Connect')
         # self.register_command('disconnect', PAMDisconnect(), 'Disconnect')
         self.default_verb = 'list'
-
-
-class PAMTunnelCommand(GroupCommand):
-
-    def __init__(self):
-        super(PAMTunnelCommand, self).__init__()
-        self.register_command('start', PAMTunnelStartCommand(), 'Start Tunnel', 's')
-        self.register_command('list', PAMTunnelListCommand(), 'List all Tunnels', 'l')
-        self.register_command('stop', PAMTunnelStopCommand(), 'Stop Tunnel to the server', 'x')
-        self.register_command('tail', PAMTunnelTailCommand(), 'View Tunnel Log', 't')
-        self.register_command('edit', PAMTunnelEditCommand(), 'Edit Tunnel settings', 'e')
-        self.default_verb = 'list'
-
-
-class PAMConnectionCommand(GroupCommand):
-
-    def __init__(self):
-        super(PAMConnectionCommand, self).__init__()
-        # self.register_command('start', PAMConnectionStartCommand(), 'Start Connection', 's')
-        # self.register_command('stop', PAMConnectionStopCommand(), 'Stop Connection', 'x')
-        self.register_command('edit', PAMConnectionEditCommand(), 'Edit Connection settings', 'e')
-        self.default_verb = 'edit'
 
 
 class PAMConfigurationsCommand(GroupCommand):
@@ -2629,1231 +2598,3 @@ class PAMCreateGatewayCommand(Command):
             print('-----------------------------------------------')
             print(bcolors.OKGREEN + one_time_token + bcolors.ENDC)
             print('-----------------------------------------------')
-
-
-# TUNNELING
-class PAMTunnelListCommand(Command):
-    pam_cmd_parser = argparse.ArgumentParser(prog='pam tunnel list')
-
-    def get_parser(self):
-        return PAMTunnelListCommand.pam_cmd_parser
-
-    def execute(self, params, **kwargs):
-        def gather_tabel_row_data(thread):
-            # {"thread": t, "host": host, "port": port, "started": datetime.now(),
-            r_row = []
-            run_time = None
-            hours = 0
-            minutes = 0
-            seconds = 0
-
-            entrance = thread.get('entrance')
-            if entrance is not None:
-                r_row.append(f"{bcolors.OKBLUE}{entrance.pc.endpoint_name}{bcolors.ENDC}")
-            else:
-                r_row.append(f"{bcolors.WARNING}Connecting..{bcolors.ENDC}")
-
-            r_row.append(f"{thread.get('host', '')}")
-
-            if entrance is not None and entrance.print_ready_event.is_set():
-                if thread.get('started'):
-                    run_time = datetime.now() - thread.get('started')
-                    hours, remainder = divmod(run_time.seconds, 3600)
-                    minutes, seconds = divmod(remainder, 60)
-
-                r_row.append(
-                    f"{bcolors.OKBLUE}{entrance.port}{bcolors.ENDC}"
-                )
-            else:
-                r_row.append(f"{bcolors.WARNING}Connecting...{bcolors.ENDC}")
-            r_row.append(f"{thread.get('record_uid', '')}")
-            if entrance is not None and entrance.print_ready_event.is_set():
-                text_line = ""
-                if run_time:
-                    if run_time.days == 1:
-                        text_line += f"{run_time.days} day "
-                    elif run_time.days > 1:
-                        text_line += f"{run_time.days} days "
-                    text_line += f"{hours} hr " if hours > 0 or run_time.days > 0 else ''
-                text_line += f"{minutes} min "
-                text_line += f"{seconds} sec"
-                r_row.append(text_line)
-            else:
-                r_row.append(f"{bcolors.WARNING}Connecting...{bcolors.ENDC}")
-            return r_row
-
-        if not params.tunnel_threads:
-            logging.warning(f"{bcolors.OKBLUE}No Tunnels running{bcolors.ENDC}")
-            return
-
-        table = []
-        headers = ['Tunnel ID', 'Host', 'Port', 'Record UID', 'Up Time']
-
-        for i, convo_id in enumerate(params.tunnel_threads):
-            row = gather_tabel_row_data(params.tunnel_threads[convo_id])
-            if row:
-                table.append(row)
-
-        dump_report_data(table, headers, fmt='table', filename="", row_number=False, column_width=None)
-
-
-def clean_up_tunnel(params, convo_id):
-    tunnel_data = None
-    index = None
-    for i, co in enumerate(params.tunnel_threads):
-        tmp_entrance = params.tunnel_threads[co].get('entrance', {})
-        if tmp_entrance and tmp_entrance.pc.endpoint_name == convo_id:
-            tunnel_data = params.tunnel_threads[co]
-            index = i
-            break
-    if tunnel_data:
-        kill_server_event = tunnel_data.get("kill_server_event")
-        if kill_server_event:
-            if not kill_server_event.is_set():
-                kill_server_event.set()
-            # whatever the read timeout is, wait for 2 seconds more
-            time.sleep(READ_TIMEOUT + 2)
-        p = tunnel_data.get("process", None)
-        if p and p.is_alive():
-            p.join()
-        if params.tunnel_threads.get(index):
-            del params.tunnel_threads[index]
-        if params.tunnel_threads_queue.get(index):
-            del params.tunnel_threads_queue[index]
-    else:
-        if params.debug:
-            print(f"{bcolors.WARNING}No tunnel data found to remove for {convo_id}{bcolors.ENDC}")
-
-
-class PAMTunnelStopCommand(Command):
-    pam_cmd_parser = argparse.ArgumentParser(prog='pam tunnel stop')
-    pam_cmd_parser.add_argument('uid', type=str, action='store', help='The Tunnel UID or Record UID')
-
-    def get_parser(self):
-        return PAMTunnelStopCommand.pam_cmd_parser
-
-    def execute(self, params, **kwargs):
-        convo_id = kwargs.get('uid')
-        if not convo_id:
-            raise CommandError('tunnel stop', '"uid" argument is required')
-        tunnel_data = []
-        for co in params.tunnel_threads:
-            tmp_entrance = params.tunnel_threads[co].get('entrance', {})
-            if tmp_entrance and tmp_entrance.pc.endpoint_name == convo_id:
-                tunnel_data.append(tmp_entrance)
-            elif tmp_entrance and tmp_entrance.pc.record_uid == convo_id:
-                tunnel_data.append(tmp_entrance)
-        if not tunnel_data:
-            raise CommandError('tunnel stop', f"No tunnel data to remove found for {convo_id}")
-        for co in tunnel_data:
-            clean_up_tunnel(params, co.pc.endpoint_name)
-        clean_up_tunnel(params, convo_id)
-
-        return
-
-
-class PAMTunnelTailCommand(Command):
-    pam_cmd_parser = argparse.ArgumentParser(prog='pam tunnel tail')
-    pam_cmd_parser.add_argument('uid', type=str, action='store', help='The Tunnel UID')
-
-    def get_parser(self):
-        return PAMTunnelTailCommand.pam_cmd_parser
-
-    def execute(self, params, **kwargs):
-        convo_id = kwargs.get('uid')
-        if not convo_id:
-            raise CommandError('tunnel tail', '"uid" argument is required')
-        tunnel_data = None
-        index = None
-        for i, co in enumerate(params.tunnel_threads):
-            tmp_entrance = params.tunnel_threads[co].get('entrance', {})
-            if tmp_entrance and tmp_entrance.pc.endpoint_name == convo_id:
-                tunnel_data = tmp_entrance
-                index = i
-                break
-        if not tunnel_data:
-            raise CommandError('tunnel tail', f"Tunnel UID {convo_id} not found")
-
-        log_queue = params.tunnel_threads_queue.get(index)
-
-        logger_level = logging.getLogger().getEffectiveLevel()
-        aio_log_level = logging.getLogger('aiortc').getEffectiveLevel()
-
-        logging.getLogger('aiortc').setLevel(logging.DEBUG)
-        logging.getLogger('aioice').setLevel(logging.DEBUG)
-        logging.getLogger(tunnel_data.pc.endpoint_name).setLevel(logging.DEBUG)
-
-        if log_queue:
-            try:
-                while True:
-                    while not log_queue.empty():
-                        print(f'    {bcolors.OKBLUE}{log_queue.get()}{bcolors.ENDC}')
-            except KeyboardInterrupt:
-                print(f'    {bcolors.WARNING}Exiting tail command{bcolors.ENDC}')
-                return
-
-            except Exception as e:
-                print(f'    {bcolors.WARNING}Exiting due to exception: {e}{bcolors.ENDC}')
-                return
-            finally:
-                logging.getLogger('aiortc').setLevel(aio_log_level)
-                logging.getLogger('aioice').setLevel(aio_log_level)
-                logging.getLogger(tunnel_data.pc.endpoint_name).setLevel(logger_level)
-        else:
-            print(f'    {bcolors.FAIL}Invalid conversation ID{bcolors.ENDC}')
-            return
-
-
-class SocketNotConnectedException(Exception):
-    pass
-
-
-class PAMTunnelEditCommand(Command):
-    pam_cmd_parser = argparse.ArgumentParser(prog='pam tunnel edit')
-    pam_cmd_parser.add_argument('record', type=str, action='store', help='The record path or UID of the PAM '
-                                                                      'resource record with network information to use '
-                                                                      'for tunneling')
-    pam_cmd_parser.add_argument('--configuration', '-c', required=False, dest='config', action='store',
-                                help='The PAM Configuration UID or path to use for tunneling. '
-                                     'Use command `pam config list` to view available PAM Configurations.')
-    pam_cmd_parser.add_argument('--enable-tunneling', '-et', required=False, dest='enable_tunneling', action='store_true',
-                                help='Enable tunneling on the record')
-    pam_cmd_parser.add_argument('--tunneling-override-port', '-top', required=False, dest='tunneling_override_port',
-                                action='store', help='Port to use for tunneling. If not provided, '
-                                                     'the port from the record will be used.')
-    pam_cmd_parser.add_argument('--disable-tunneling', '-dt', required=False, dest='disable_tunneling',
-                                action='store_true', help='Disable tunneling on the record')
-    pam_cmd_parser.add_argument('--remove-tunneling-override-port', '-rtop', required=False,
-                                dest='remove_tunneling_override_port', action='store_true',
-                                help='Remove tunneling override port')
-
-    def get_parser(self):
-        return PAMTunnelEditCommand.pam_cmd_parser
-
-    def execute(self, params, **kwargs):
-        tunneling_override_port = kwargs.get('tunneling_override_port')
-
-        if ((kwargs.get('enable_tunneling') and kwargs.get('disable_tunneling')) or
-                (kwargs.get('enable_rotation') and kwargs.get('disable_rotation')) or
-                (kwargs.get('tunneling-override-port') and kwargs.get('remove_tunneling_override_port'))):
-            raise CommandError('pam-config-edit', 'Cannot enable and disable the same feature at the same time')
-
-        # First check if enabled is true then check if disabled is true. if not then set it to None
-        _tunneling = True if kwargs.get('enable_tunneling') else False if  kwargs.get('disable_tunneling') else None
-        _remove_tunneling_override_port = kwargs.get('remove_tunneling_override_port')
-
-        if tunneling_override_port:
-            try:
-                tunneling_override_port = int(tunneling_override_port)
-            except ValueError:
-                raise CommandError('tunnel edit', 'tunneling-override-port must be an integer')
-
-        record_name = kwargs.get('record')
-        if not record_name:
-            raise CommandError('pam tunnel edit', '"record" parameter is required.')
-        record = RecordMixin.resolve_single_record(params, record_name)
-        if not record:
-            raise CommandError('pam tunnel edit', f'{bcolors.FAIL}Record \"{record_name}\" not found.{bcolors.ENDC}')
-        if not isinstance(record, vault.TypedRecord):
-            raise CommandError('pam tunnel edit', f'Record \"{record_name}\" can not be edited.')
-
-        # config parameter is optional and may be (auto)resolved from PAM record
-        config_name = kwargs.get('config', None)
-        cfg_rec = RecordMixin.resolve_single_record(params, config_name)
-        if not cfg_rec and record.version == 6:
-            cfg_rec = record  # trying to edit PAM Config itself
-        config_uid = cfg_rec.record_uid if cfg_rec else None
-
-        record_uid = record.record_uid
-        record_type = record.record_type
-        if record_type not in ("pamMachine pamDatabase pamDirectory pamNetworkConfiguration pamAwsConfiguration "
-                               "pamRemoteBrowser pamAzureConfiguration").split():
-            raise CommandError('', f"{bcolors.FAIL}This record's type is not supported for tunnels. "
-                                   f"Tunnels are only supported on pamMachine, pamDatabase, pamDirectory, "
-                                   f"pamRemoteBrowser, pamNetworkConfiguration pamAwsConfiguration, and "
-                                   f"pamAzureConfiguration records{bcolors.ENDC}")
-
-        encrypted_session_token, encrypted_transmission_key, transmission_key = get_keeper_tokens(params)
-        if record_type in "pamNetworkConfiguration pamAwsConfiguration pamAzureConfiguration".split():
-            tmp_dag = TunnelDAG(params, encrypted_session_token, encrypted_transmission_key, record_uid, is_config=True)
-            tmp_dag.edit_tunneling_config(tunneling=_tunneling)
-            tmp_dag.print_tunneling_config(record_uid, None)
-        else:
-            traffic_encryption_key = record.get_typed_field('trafficEncryptionSeed')
-            # Generate a 256-bit (32-byte) random seed
-            seed = os.urandom(32)
-            dirty = False
-            if not traffic_encryption_key or not traffic_encryption_key.value:
-                base64_seed = bytes_to_base64(seed)
-                record_seed = vault.TypedField.new_field('trafficEncryptionSeed', base64_seed, "")
-                # if field is present update in-place, if in rec definition add to fields[] else custom[]
-                record_types_with_seed = ("pamDatabase", "pamDirectory", "pamMachine", "pamRemoteBrowser")
-                if traffic_encryption_key:
-                    traffic_encryption_key.value = [base64_seed]
-                elif record.get_record_type() in record_types_with_seed:
-                    record.fields.append(record_seed)  # DU-469
-                else:
-                    record.custom.append(record_seed)
-                dirty = True
-            if dirty:
-                record_management.update_record(params, record)
-                api.sync_down(params)
-
-                traffic_encryption_key = record.get_typed_field('trafficEncryptionSeed')
-                if not traffic_encryption_key:
-                    raise CommandError('', f"{bcolors.FAIL}Unable to add Seed to record {record_uid}. "
-                                       f"Please make sure you have edit rights to record {record_uid} {bcolors.ENDC}")
-            dirty = False
-
-            existing_config_uid = get_config_uid(params, encrypted_session_token, encrypted_transmission_key, record_uid)
-
-            tmp_dag = TunnelDAG(params, encrypted_session_token, encrypted_transmission_key, config_uid)
-            old_dag = TunnelDAG(params, encrypted_session_token, encrypted_transmission_key, existing_config_uid)
-
-            if config_uid and existing_config_uid != config_uid:
-                old_dag.remove_from_dag(record_uid)
-                tmp_dag.link_resource_to_config(record_uid)
-
-            if tmp_dag is None or not tmp_dag.linking_dag.has_graph:
-                raise CommandError('', f"{bcolors.FAIL}No PAM Configuration UID set. "
-                                   f"This must be set or supplied for tunneling to work. This can be done by adding "
-                                   f"{bcolors.OKBLUE}' --config [ConfigUID] "
-                                   f" {bcolors.FAIL}The ConfigUID can be found by running "
-                                   f"{bcolors.OKBLUE}'pam config list'{bcolors.ENDC}")
-
-            if not tmp_dag.check_tunneling_enabled_config(enable_tunneling=_tunneling):
-                tmp_dag.print_tunneling_config(config_uid, None)
-                command = f"{bcolors.OKBLUE}'pam tunnel edit {config_uid}"
-                if _tunneling and not tmp_dag.check_tunneling_enabled_config(
-                        enable_tunneling=_tunneling):
-                    command += f" --enable-tunneling" if _tunneling else ""
-
-                print(f"{bcolors.FAIL}The settings are denied by PAM Configuration: {config_uid}. "
-                      f"Please enable settings for the configuration by running\n"
-                      f"{command}'{bcolors.ENDC}")
-                return
-
-            if not tmp_dag.is_tunneling_config_set_up(record_uid):
-                tmp_dag.link_resource_to_config(record_uid)
-
-            pam_settings = record.get_typed_field('pamSettings')
-            if not pam_settings:
-                pre_settings = {"connection": {}, "portForward": {}}
-                if _tunneling and tunneling_override_port:
-                    pre_settings["portForward"]["port"] = tunneling_override_port
-                if pre_settings:
-                    pam_settings = vault.TypedField.new_field('pamSettings', pre_settings, "")
-                    # TODO follow template
-                    record.custom.append(pam_settings)
-                    dirty = True
-            else:
-                if not tmp_dag.is_tunneling_config_set_up(record_uid):
-                    tmp_dag.link_resource_to_config(record_uid)
-                if not pam_settings.value:
-                    pam_settings.value.append({"connection": {}, "portForward": {}})
-                if _tunneling and tunneling_override_port:
-                    pam_settings.value[0]['portForward']['port'] = tunneling_override_port
-                    dirty = True
-
-                if _remove_tunneling_override_port and pam_settings.value[0]['portForward'].get('port'):
-                    pam_settings.value[0]['portForward'].pop('port')
-                    dirty = True
-            if not tmp_dag.is_tunneling_config_set_up(record_uid):
-                print(f"{bcolors.FAIL}No PAM Configuration UID set. This must be set for tunneling to work. "
-                      f"This can be done by running "
-                      f"{bcolors.OKBLUE}'pam tunnel edit {record_uid} --config [ConfigUID] --enable-tunneling' "
-                      f"{bcolors.FAIL}The ConfigUID can be found by running "
-                      f"{bcolors.OKBLUE}'pam config list'{bcolors.ENDC}")
-                return
-            allowed_settings_name = "allowedSettings"
-            if record.record_type == "pamRemoteBrowser":
-                allowed_settings_name = "pamRemoteBrowserSettings"
-
-            if _tunneling is not None and tmp_dag.check_if_resource_allowed(record_uid, "portForwards") != _tunneling:
-                dirty = True
-
-            if dirty:
-                tmp_dag.set_resource_allowed(resource_uid=record_uid, tunneling=_tunneling, allowed_settings_name=allowed_settings_name)
-
-            # Print out the tunnel settings
-            if not kwargs.get('silent'):
-                tmp_dag.print_tunneling_config(record_uid, record.get_typed_field('pamSettings'), config_uid)
-
-
-class PAMTunnelStartCommand(Command):
-    pam_cmd_parser = argparse.ArgumentParser(prog='pam tunnel start')
-    pam_cmd_parser.add_argument('uid', type=str, action='store', help='The Record UID of the PAM resource '
-                                                                      'record with network information to use for '
-                                                                      'tunneling')
-    pam_cmd_parser.add_argument('--host', '-o', required=False, dest='host', action='store',
-                                default="127.0.0.1",
-                                help='The address on which the server will be accepting connections. It could be an '
-                                     'IP address or a hostname. '
-                                     'Ex. set to 127.0.0.1 as default so only connections from the same machine will be'
-                                     ' accepted.')
-    pam_cmd_parser.add_argument('--port', '-p', required=False, dest='port', action='store',
-                                type=int, default=0,
-                                help='The port number on which the server will be listening for incoming connections. '
-                                     'If not set, random open port on the machine will be used.')
-
-    def get_parser(self):
-        return PAMTunnelStartCommand.pam_cmd_parser
-
-    class QueueHandler(logging.Handler):
-        # Custom logging handler that will put messages into a queue of the tunnel threads
-        def __init__(self, log_queue):
-            super().__init__()
-            self.log_queue = log_queue
-
-        def emit(self, record):
-            log_entry = self.format(record)
-            try:
-                self.log_queue.put_nowait(log_entry)
-            except queue.Full:
-                # If the queue is full, remove the oldest (first) item
-                self.log_queue.get_nowait()
-                # Then add the new log entry
-                self.log_queue.put_nowait(log_entry)
-
-    def setup_logging(self, convo_id, log_queue, logging_level):
-        logger = logging.getLogger(convo_id)
-        logger.setLevel(logging_level)
-        logger.propagate = False
-        queue_handler = self.QueueHandler(log_queue)
-        logger.addHandler(queue_handler)
-        logger.debug("Logging setup complete.")
-        return logger
-
-    async def connect(self, params, record_uid, gateway_uid, convo_num, host, port,
-                      log_queue, seed, target_host, target_port, socks):
-
-        # Setup custom logging to put logs into log_queue
-        logger = self.setup_logging(str(convo_num), log_queue, logging.getLogger().getEffectiveLevel())
-
-        print(f"{bcolors.HIGHINTENSITYWHITE}Establishing tunnel between Commander and Gateway. Please wait..."
-              f"{bcolors.ENDC}")
-
-        # Symmetric key
-        """
-        Generate a 256-bit (32-byte) random seed
-        seed = os.urandom(32)
-        """
-        if isinstance(seed, str):
-            seed = base64_to_bytes(seed)
-        # Generate a 128-bit (16-byte) random nonce
-        nonce = os.urandom(MAIN_NONCE_LENGTH)
-        # Derive the encryption key using HKDF
-        hkdf = HKDF(
-            algorithm=hashes.SHA256(),
-            length=SYMMETRIC_KEY_LENGTH,  # 256-bit key
-            salt=nonce,
-            info=b"KEEPER_TUNNEL_ENCRYPT_AES_GCM_128",
-            backend=default_backend()
-        ).derive(seed)
-        symmetric_key = AESGCM(hkdf)
-
-        # Set up the pc
-        print_ready_event = asyncio.Event()
-        kill_server_event = asyncio.Event()
-        pc = WebRTCConnection(params=params, record_uid=record_uid, gateway_uid=gateway_uid, symmetric_key=symmetric_key,
-                              print_ready_event=print_ready_event, kill_server_event=kill_server_event,
-                              logger=logger, server=params.server)
-
-        try:
-            await pc.signal_channel('start', bytes_to_base64(nonce))
-        except Exception as e:
-            raise CommandError('Tunnel Start', f"{e}")
-
-        logger.debug("starting private tunnel")
-
-        if socks:
-            private_tunnel = SOCKS5Server(host=host, port=port, pc=pc, print_ready_event=print_ready_event,
-                                          logger=logger,
-                                          connect_task=params.tunnel_threads[convo_num].get("connect_task", None),
-                                          kill_server_event=kill_server_event, target_host=target_host,
-                                          target_port=target_port)
-        else:
-            private_tunnel = TunnelEntrance(host=host, port=port, pc=pc, print_ready_event=print_ready_event,
-                                            logger=logger,
-                                            connect_task=params.tunnel_threads[convo_num].get("connect_task", None),
-                                            kill_server_event=kill_server_event, target_host=target_host,
-                                            target_port=target_port)
-
-        t1 = asyncio.create_task(private_tunnel.start_server())
-        params.tunnel_threads[convo_num].update({"server": t1, "entrance": private_tunnel,
-                                                 "kill_server_event": kill_server_event})
-
-        logger.debug("--> START LISTENING FOR MESSAGES FROM GATEWAY --------")
-        try:
-            await asyncio.gather(t1, private_tunnel.reader_task)
-        except asyncio.CancelledError:
-            pass
-        finally:
-            logger.debug("--> STOP LISTENING FOR MESSAGES FROM GATEWAY --------")
-
-    def pre_connect(self, params, record_uid, gateway_uid, convo_num, host, port,
-                    seed, target_host, target_port, socks):
-        tunnel_name = f"{convo_num}"
-
-        def custom_exception_handler(_loop, context):
-            # Check if the exception is present in the context
-            if "exception" in context:
-                exception = context["exception"]
-                if isinstance(exception, ConnectionError):
-                    # Handle only ConnectionError
-                    logging.debug(f"Caught ConnectionError in asyncio: {exception}")
-            else:
-                # Log the default message if no exception is found
-                logging.error(context["message"])
-
-        loop = None
-        try:
-            # Create a new asyncio event loop and set the custom exception handler
-            loop = asyncio.new_event_loop()
-            params.tunnel_threads[convo_num].update({"loop": loop})
-            asyncio.set_event_loop(loop)
-            loop.set_exception_handler(custom_exception_handler)
-            output_queue = queue.Queue(maxsize=500)
-            params.tunnel_threads_queue[convo_num] = output_queue
-            # Create a Task from the coroutine
-            connect_task = loop.create_task(
-                self.connect(
-                    params=params,
-                    record_uid=record_uid,
-                    gateway_uid=gateway_uid,
-                    convo_num=convo_num,
-                    host=host,
-                    port=port,
-                    log_queue=output_queue,
-                    seed=seed,
-                    target_host=target_host,
-                    target_port=target_port,
-                    socks=socks
-                )
-            )
-            params.tunnel_threads[convo_num].update({"connect_task": connect_task})
-            try:
-                # Run the task until it is complete
-                loop.run_until_complete(connect_task)
-            except asyncio.CancelledError:
-                pass
-        except SocketNotConnectedException as es:
-            print(f"{bcolors.FAIL}Socket not connected exception in connection {tunnel_name}: {es}{bcolors.ENDC}")
-        except KeyboardInterrupt:
-            print(f"{bcolors.OKBLUE}Exiting: connection {tunnel_name}{bcolors.ENDC}")
-        except CommandError as ce:
-            print(f"{bcolors.FAIL}{ce}{bcolors.ENDC}")
-        except Exception as e:
-            print(f"{bcolors.FAIL}An exception occurred in connection {tunnel_name}: {e}{bcolors.ENDC}")
-        finally:
-            if loop:
-                try:
-                    tunnel_data = params.tunnel_threads.get(convo_num, None)
-                    co_entrance = tunnel_data.get('entrance')
-                    if co_entrance:
-                        tunnel_name = co_entrance.pc.endpoint_name
-                    if not tunnel_data:
-                        logging.debug(f"{bcolors.WARNING}No tunnel data found for {tunnel_name}{bcolors.ENDC}")
-                        return
-
-                    if convo_num in params.tunnel_threads_queue:
-                        del params.tunnel_threads_queue[convo_num]
-
-                    entrance = tunnel_data.get("entrance", None)
-                    if entrance:
-                        loop.run_until_complete(entrance.stop_server(CloseConnectionReasons.ConnectionFailed))
-
-                    del params.tunnel_threads[convo_num]
-                    logging.debug(f"Cleaned up data for {tunnel_name}")
-
-                    try:
-                        for task in asyncio.all_tasks(loop):
-                            task.cancel()
-                        loop.stop()
-                        loop.close()
-                        logging.debug(f"{tunnel_name} Loop cleaned up")
-                    except Exception as e:
-                        logging.debug(f"{bcolors.WARNING}Exception while stopping event loop: {e}{bcolors.ENDC}")
-                except Exception as e:
-                    print(
-                        f"{bcolors.FAIL}An exception occurred in pre_connect for connection {tunnel_name}: {e}"
-                        f"{bcolors.ENDC}")
-                finally:
-                    clean_up_tunnel(params, convo_num)
-                    print(f"{bcolors.OKBLUE}Tunnel {tunnel_name} closed.{bcolors.ENDC}")
-
-    def execute(self, params, **kwargs):
-        # https://pypi.org/project/aiortc/
-        # aiortc Requires: Python >=3.8
-        from_version = [3, 8, 0]   # including
-        to_version = [3, 13, 0]    # excluding
-        major_version = sys.version_info.major
-        minor_version = sys.version_info.minor
-        micro_version = sys.version_info.micro
-
-        if (major_version, minor_version, micro_version) < (from_version[0], from_version[1], from_version[2]):
-            print(f"{bcolors.FAIL}This command requires Python {from_version[0]}.{from_version[1]}.{from_version[2]} or higher. "
-                  f"You are using {major_version}.{minor_version}.{micro_version}.{bcolors.ENDC}")
-            return
-        if (major_version, minor_version, micro_version) >= (to_version[0], to_version[1], to_version[2]):
-            print(f"{bcolors.FAIL}This command is compatible with Python versions below {to_version[0]}.{to_version[1]}.{to_version[2]} "
-                  f"(Current Python version: {major_version}.{minor_version}.{micro_version}){bcolors.ENDC}")
-            return
-
-        record_uid = kwargs.get('uid')
-        convo_num = len(params.tunnel_threads)
-        params.tunnel_threads[convo_num] = {}
-        host = kwargs.get('host')
-        port = kwargs.get('port')
-        if port is not None and port > 0:
-            try:
-                port = find_open_port(tried_ports=[], preferred_port=port, host=host)
-            except CommandError as e:
-                print(f"{bcolors.FAIL}{e}{bcolors.ENDC}")
-                del params.tunnel_threads[convo_num]
-                return
-        else:
-            port = find_open_port(tried_ports=[], host=host)
-            if port is None:
-                print(f"{bcolors.FAIL}Could not find open port to use for tunnel{bcolors.ENDC}")
-                del params.tunnel_threads[convo_num]
-                return
-
-        api.sync_down(params)
-        record = vault.KeeperRecord.load(params, record_uid)
-        if not isinstance(record, vault.TypedRecord):
-            print(f"{bcolors.FAIL}Record {record_uid} not found.{bcolors.ENDC}")
-            return
-
-        pam_settings = record.get_typed_field('pamSettings')
-        if not pam_settings:
-            print(f"{bcolors.FAIL}PAM Settings not configured for record {record_uid}'.{bcolors.ENDC}")
-            print(f"{bcolors.WARNING}This is done by running {bcolors.OKBLUE}'pam tunnel edit {record_uid} "
-                  f"--enable-tunneling --config [ConfigUID]'"
-                  f"{bcolors.WARNING} The ConfigUID can be found by running"
-                  f"{bcolors.OKBLUE} 'pam config list'{bcolors.ENDC}.")
-            return
-
-        # SOCKS5 Proxy uses this to determine what connection to use for the tunnel
-        target = record.get_typed_field('pamHostname')
-        if not target:
-            print(f"{bcolors.FAIL}Hostname not found for record {record_uid}.{bcolors.ENDC}")
-            return
-        target_host = target.get_default_value().get('hostName', None)
-        target_port = target.get_default_value().get('port', None)
-        if not target_host:
-            print(f"{bcolors.FAIL}Host not found for record {record_uid}.{bcolors.ENDC}")
-            return
-        if not target_port:
-            print(f"{bcolors.FAIL}Port not found for record {record_uid}.{bcolors.ENDC}")
-            return
-
-        # IP or a CIDR subnet.
-        allowed_hosts = record.get_typed_field('multiline', 'Allowed Hosts')
-
-        allowed_ports = record.get_typed_field('multiline', 'Allowed Ports')
-        socks = False
-        if allowed_hosts or allowed_ports:
-            socks = True
-
-        client_private_seed = record.get_typed_field('trafficEncryptionSeed')
-        if not client_private_seed:
-            print(f"{bcolors.FAIL}Traffic Encryption Seed not found for record {record_uid}.{bcolors.ENDC}")
-            return
-        base64_seed = client_private_seed.get_default_value(str).encode('utf-8')
-        seed = base64_to_bytes(base64_seed)
-
-        # gateway = kwargs.get('gateway_uid')  # type: Optional[str]
-        # if gateway:
-        #     gateways = gateway_helper.get_all_gateways(params)
-        #     gateway_uid = next((utils.base64_url_encode(x.controllerUid) for x in gateways
-        #                         if utils.base64_url_encode(x.controllerUid) == gateway
-        #                         or x.controllerName.casefold() == gateway.casefold()), None)
-        gateway_uid = self.get_gateway_uid_from_record(params, record_uid)
-        if not gateway_uid:
-            print(f"{bcolors.FAIL}Gateway not found for record {record_uid}.{bcolors.ENDC}")
-            return
-
-        t = threading.Thread(target=self.pre_connect, args=(params, record_uid, gateway_uid, convo_num,
-                                                            host, port, seed, target_host, target_port, socks)
-                             )
-
-        # Setting the thread as a daemon thread
-        t.daemon = True
-        t.start()
-
-        if not params.tunnel_threads.get(convo_num):
-            params.tunnel_threads[convo_num] = {"thread": t, "host": host, "port": port,
-                                                "started": datetime.now(), "record_uid": record_uid}
-        else:
-            entrance = params.tunnel_threads[convo_num].get("entrance", None)
-            if entrance is not None:
-                endpoint_name = entrance.pc.endpoint_name
-            params.tunnel_threads[convo_num].update({"thread": t, "host": host, "port": port,
-                                                    "started": datetime.now(), "record_uid": record_uid})
-        count = 0
-        wait_time = 120
-        entrance = None
-        while count < wait_time:
-            if params.tunnel_threads.get(convo_num):
-                entrance = params.tunnel_threads[convo_num].get("entrance", None)
-                if entrance:
-                    break
-            else:
-                break
-            count += .1
-            time.sleep(.1)
-
-        def print_fail(con_num):
-            con_name = ''
-            con_entrance = None
-            if con_num in params.tunnel_threads:
-                con_entrance = params.tunnel_threads[con_num].get("entrance", None)
-            fail_dynamic_length = len("| Endpoint ") + len(" failed to start..")
-            if con_entrance:
-                con_name = con_entrance.pc.endpoint_name
-                fail_dynamic_length = len("| Endpoint ") + len(con_name) + len(" failed to start..")
-
-                clean_up_tunnel(params, con_entrance.pc.endpoint_name)
-                time.sleep(.5)
-            # Dashed line adjusted to the length of the middle line
-            fail_dashed_line = '+' + '-' * fail_dynamic_length + '+'
-            print(f'\n{bcolors.FAIL}{fail_dashed_line}{bcolors.ENDC}')
-            print(f'{bcolors.FAIL}| Endpoint {bcolors.ENDC}{con_name}{bcolors.FAIL} failed to start..{bcolors.ENDC}')
-            print(f'{bcolors.FAIL}{fail_dashed_line}{bcolors.ENDC}\n')
-
-        if entrance is not None:
-            while not entrance.print_ready_event.is_set() and count < wait_time * 2:
-                count += .1
-                time.sleep(.1)
-                if entrance.kill_server_event.is_set():
-                    break
-
-            if entrance.print_ready_event.is_set():
-                # Sleep a little bit to print out last
-                time.sleep(.5)
-                host = host + ":" if host else ''
-                # Total length of the dynamic parts (endpoint name, host, and port)
-                dynamic_length = \
-                    (len("| Endpoint : Listening on: ") +
-                     len(entrance.pc.endpoint_name) +
-                     len(host) +
-                     len(str(entrance.port)))
-
-                # Dashed line adjusted to the length of the middle line
-                dashed_line = '+' + '-' * dynamic_length + '+'
-
-                endpoint_name = entrance.pc.endpoint_name
-
-                # Print statements
-                print(f'\n{bcolors.OKGREEN}{dashed_line}{bcolors.ENDC}')
-                print(
-                    f'{bcolors.OKGREEN}| Endpoint {bcolors.ENDC}{bcolors.OKBLUE}{endpoint_name}{bcolors.ENDC}'
-                    f'{bcolors.OKGREEN}: Listening on: {bcolors.ENDC}'
-                    f'{bcolors.BOLD}{bcolors.OKBLUE}{host}{entrance.port}{bcolors.ENDC}{bcolors.OKGREEN} |{bcolors.ENDC}')
-                print(f'{bcolors.OKGREEN}{dashed_line}{bcolors.ENDC}')
-                print(
-                    f'{bcolors.OKGREEN}View all open tunnels   : {bcolors.ENDC}{bcolors.OKBLUE}pam tunnel list{bcolors.ENDC}')
-                print(f'{bcolors.OKGREEN}Tail logs on open tunnel: {bcolors.ENDC}'
-                      f'{bcolors.OKBLUE}pam tunnel tail ' +
-                      (f'-- ' if endpoint_name[0] == '-' else '') +
-                      f'{endpoint_name}{bcolors.ENDC}')
-                print(f'{bcolors.OKGREEN}Stop a tunnel           : {bcolors.ENDC}'
-                      f'{bcolors.OKBLUE}pam tunnel stop ' +
-                      (f'-- ' if endpoint_name[0] == '-' else '') +
-                      f'{endpoint_name}{bcolors.ENDC}\n')
-            else:
-                print_fail(convo_num)
-        else:
-            print_fail(convo_num)
-
-    def get_config_uid_from_record(self, params, record_uid):
-        record = vault.KeeperRecord.load(params, record_uid)
-        if not isinstance(record, vault.TypedRecord):
-            raise CommandError('', f"{bcolors.FAIL}Record {record_uid} not found.{bcolors.ENDC}")
-        record_type = record.record_type
-        if record_type not in ("pamMachine pamDatabase pamDirectory pamRemoteBrowser").split():
-            raise CommandError('', f"{bcolors.FAIL}This record's type is not supported for tunnels. "
-                                f"Tunnels are only supported on pamMachine, pamDatabase, pamDirectory, "
-                                f"and pamRemoteBrowser records{bcolors.ENDC}")
-
-        encrypted_session_token, encrypted_transmission_key, transmission_key = get_keeper_tokens(params)
-        existing_config_uid = get_config_uid(params, encrypted_session_token, encrypted_transmission_key, record_uid)
-        return existing_config_uid
-
-    def get_gateway_uid_from_record(self, params, record_uid):
-        gateway_uid = ''
-        pam_config_uid = self.get_config_uid_from_record(params, record_uid)
-        if pam_config_uid:
-            record = vault.KeeperRecord.load(params, pam_config_uid)
-            if record:
-                field = record.get_typed_field('pamResources')
-                value = field.get_default_value(dict)
-                if value:
-                    gateway_uid = value.get('controllerUid', '') or ''
-
-        return gateway_uid
-
-
-class PAMConnectionEditCommand(Command):
-    choices = ['on', 'off', 'default']
-    protocols = ['', 'http', 'kubernetes', 'mysql', 'postgresql', 'rdp', 'sql-server', 'ssh', 'telnet', 'vnc']
-    parser = argparse.ArgumentParser(prog='pam connection edit')
-    parser.add_argument('record', type=str, action='store', help='The record UID or path of the PAM '
-                        'resource record with network information to use for connections')
-    parser.add_argument('--configuration', '-c', required=False, dest='config', action='store',
-                        help='The PAM Configuration UID or path to use for connections. '
-                        'Use command `pam config list` to view available PAM Configurations.')
-    parser.add_argument('--admin-user', '-a', required=False, dest='admin', action='store',
-					help='The record path or UID of the PAM User record to configure the admin '
-                    'credential on the PAM Resource')
-    parser.add_argument('--protocol', '-p', dest='protocol', choices=protocols,
-                        help='Set connection protocol')
-    parser.add_argument('--connections', '-cn', dest='connections', choices=choices,
-                        help='Set connections permissions')
-    parser.add_argument('--connections-recording', '-cr', dest='recording', choices=choices,
-                        help='Set recording connections permissions for the resource')
-    parser.add_argument('--typescript-recording', '-tr', dest='typescriptrecording', choices=choices,
-                        help='Set TypeScript recording permissions for the resource')
-    parser.add_argument('--connections-override-port', '-cop', required=False, dest='connections_override_port',
-                        action='store', help='Port to use for connections. If not provided, '
-                        'the port from the record will be used.')
-    parser.add_argument('--silent', '-s', required=False, dest='silent', action='store_true',
-					help='Silent mode - don\'t print PAM User, PAM Config etc.')
-
-    def get_parser(self):
-        return PAMConnectionEditCommand.parser
-
-    def execute(self, params, **kwargs):
-        connection_override_port = kwargs.get('connections_override_port', None)
-
-        # Convert on/off/default to True/False/None
-        _connections = TunnelDAG._convert_allowed_setting(kwargs.get('connections', None))
-        _recording = TunnelDAG._convert_allowed_setting(kwargs.get('recording', None))
-        _typescript_recording = TunnelDAG._convert_allowed_setting(kwargs.get('typescriptrecording', None))
-
-        if connection_override_port:
-            try:
-                connection_override_port = int(connection_override_port)
-            except ValueError:
-                raise CommandError('connection edit', '--connections-override-port must be an integer')
-
-        record_name = kwargs.get('record')
-        if not record_name:
-            raise CommandError('pam connection edit', 'Record parameter is required.')
-        record = RecordMixin.resolve_single_record(params, record_name)
-        if not record:
-            raise CommandError('pam connection edit', f'{bcolors.FAIL}Record \"{record_name}\" not found.{bcolors.ENDC}')
-        if not isinstance(record, vault.TypedRecord):
-            raise CommandError('pam connection edit', f'Record \"{record_name}\" can not be edited.')
-
-        # config parameter is optional and may be (auto)resolved from PAM record
-        config_name = kwargs.get('config', None)
-        cfg_rec = RecordMixin.resolve_single_record(params, config_name)
-        if not cfg_rec and record.version == 6:
-            cfg_rec = record  # trying to edit PAM Config itself
-        config_uid = cfg_rec.record_uid if cfg_rec else None
-
-        record_uid = record.record_uid
-        record_type = record.record_type
-        if record_type not in ("pamMachine pamDatabase pamDirectory pamNetworkConfiguration pamAwsConfiguration "
-                               "pamRemoteBrowser pamAzureConfiguration").split():
-            raise CommandError('', f"{bcolors.FAIL}This record's type is not supported for connections. "
-                                   f"Connectins are only supported on pamMachine, pamDatabase, pamDirectory, "
-                                   f"pamRemoteBrowser, pamNetworkConfiguration pamAwsConfiguration, and "
-                                   f"pamAzureConfiguration records{bcolors.ENDC}")
-
-        encrypted_session_token, encrypted_transmission_key, transmission_key = get_keeper_tokens(params)
-        if record_type in "pamNetworkConfiguration pamAwsConfiguration pamAzureConfiguration".split():
-            tdag = TunnelDAG(params, encrypted_session_token, encrypted_transmission_key, record_uid, is_config=True)
-            tdag.edit_tunneling_config(connections=_connections, session_recording=_recording, typescript_recording=_typescript_recording)
-            if not kwargs.get("silent", False): tdag.print_tunneling_config(record_uid, None)
-        else:
-            traffic_encryption_key = record.get_typed_field('trafficEncryptionSeed')
-            # Generate a 256-bit (32-byte) random seed
-            seed = os.urandom(32)
-            dirty = False
-            if not traffic_encryption_key or not traffic_encryption_key.value:
-                base64_seed = bytes_to_base64(seed)
-                record_seed = vault.TypedField.new_field('trafficEncryptionSeed', base64_seed, "")
-                # if field is present update in-place, if in rec definition add to fields[] else custom[]
-                record_types_with_seed = ("pamDatabase", "pamDirectory", "pamMachine", "pamRemoteBrowser")
-                if traffic_encryption_key:
-                    traffic_encryption_key.value = [base64_seed]
-                elif record.get_record_type() in record_types_with_seed:
-                    record.fields.append(record_seed)  # DU-469
-                else:
-                    record.custom.append(record_seed)
-                dirty = True
-
-            protocol = kwargs.get("protocol", None)
-            pam_settings = record.get_typed_field('pamSettings')
-            if not pam_settings:
-                pre_settings = {"connection": {}, "portForward": {}}
-                if _connections:
-                    if connection_override_port:
-                        pre_settings["connection"]["port"] = connection_override_port
-                    if protocol:
-                        pre_settings["connection"]["protocol"] = protocol
-                elif protocol or connection_override_port:
-                    logging.warning(f'Connection override port and protocol can be set only when connections are enabled '
-                            f'with {bcolors.OKGREEN}--connections=on{bcolors.ENDC} option')
-                if pre_settings:
-                    pam_settings = vault.TypedField.new_field('pamSettings', pre_settings, "")
-                    # TODO follow template
-                    record.custom.append(pam_settings)
-                    dirty = True
-            else:
-                if not pam_settings.value:
-                    pam_settings.value.append({"connection": {}, "portForward": {}})
-                if _connections:
-                    if connection_override_port:
-                        pam_settings.value[0]["connection"]["port"] = connection_override_port
-                    elif connection_override_port is not None:  # empty string means remove port override
-                        pam_settings.value[0]["connection"].pop("port", None)
-                    if protocol:
-                        pam_settings.value[0]["connection"]["protocol"] = protocol
-                    elif protocol is not None:  # empty string means remove protocol
-                        pam_settings.value[0]["connection"].pop("protocol", None)
-                    dirty = True
-                elif protocol or connection_override_port:
-                    logging.warning(f'Connection override port and protocol can be set only when connections are enabled '
-                            f'with {bcolors.OKGREEN}--connections=on{bcolors.ENDC} option')
-            if dirty:
-                record_management.update_record(params, record)
-                api.sync_down(params)
-
-                traffic_encryption_key = record.get_typed_field('trafficEncryptionSeed')
-                if not traffic_encryption_key:
-                    raise CommandError('', f"{bcolors.FAIL}Unable to add Seed to record {record_uid}. "
-                                       f"Please make sure you have edit rights to record {record_uid} {bcolors.ENDC}")
-            dirty = False
-
-            existing_config_uid = get_config_uid(params, encrypted_session_token, encrypted_transmission_key, record_uid)
-
-            tdag = TunnelDAG(params, encrypted_session_token, encrypted_transmission_key, config_uid)
-            old_dag = TunnelDAG(params, encrypted_session_token, encrypted_transmission_key, existing_config_uid)
-
-            if config_uid and existing_config_uid != config_uid:
-                old_dag.remove_from_dag(record_uid)
-                tdag.link_resource_to_config(record_uid)
-
-            if tdag is None or not tdag.linking_dag.has_graph:
-                raise CommandError('', f"{bcolors.FAIL}No PAM Configuration UID set. "
-                                   f"This must be set or supplied for connections to work. This can be done by adding "
-                                   f"{bcolors.OKBLUE}' --config [ConfigUID] "
-                                   f" {bcolors.FAIL}The ConfigUID can be found by running "
-                                   f"{bcolors.OKBLUE}'pam config list'{bcolors.ENDC}")
-
-            if not tdag.check_tunneling_enabled_config(enable_connections=_connections,
-                                                       enable_session_recording=_recording,
-                                                       enable_typescript_recording=_typescript_recording):
-                if not kwargs.get("silent", False): tdag.print_tunneling_config(config_uid, None)
-                command = f"{bcolors.OKBLUE}'pam connection edit {config_uid}"
-                if _connections and not tdag.check_tunneling_enabled_config(enable_connections=_connections):
-                    command += f" --connections=on" if _connections else ""
-                if _recording and not tdag.check_tunneling_enabled_config(enable_session_recording=_recording):
-                    command += f" --connections-recording=on" if _recording else ""
-                if _typescript_recording and not tdag.check_tunneling_enabled_config(enable_typescript_recording=_typescript_recording):
-                    command += f" --typescript-recording=on" if _typescript_recording else ""
-
-                print(f"{bcolors.FAIL}The settings are denied by PAM Configuration: {config_uid}. "
-                      f"Please enable settings for the configuration by running\n"
-                      f"{command}'{bcolors.ENDC}")
-                return
-
-            if not tdag.is_tunneling_config_set_up(record_uid):
-                tdag.link_resource_to_config(record_uid)
-
-            if not tdag.is_tunneling_config_set_up(record_uid):
-                print(f"{bcolors.FAIL}No PAM Configuration UID set. This must be set for connections to work. "
-                      f"This can be done by running {bcolors.OKBLUE}"
-                      f"'pam connection edit {record_uid} --config [ConfigUID] --enable-connections' "
-                      f"{bcolors.FAIL}The ConfigUID can be found by running {bcolors.OKBLUE}'pam config list'{bcolors.ENDC}")
-                return
-            allowed_settings_name = "allowedSettings"
-            if record.record_type == "pamRemoteBrowser":
-                allowed_settings_name = "pamRemoteBrowserSettings"
-
-            if _connections is not None and tdag.check_if_resource_allowed(record_uid, "connections") != _connections:
-                dirty = True
-            if _recording is not None and tdag.check_if_resource_allowed(record_uid, "sessionRecording") != _recording:
-                dirty = True
-            if _typescript_recording is not None and tdag.check_if_resource_allowed(record_uid, "typescriptRecording") != _typescript_recording:
-                dirty = True
-
-            if dirty:
-                tdag.set_resource_allowed(resource_uid=record_uid,
-                                          allowed_settings_name=allowed_settings_name,
-                                          connections=kwargs.get('connections', None),
-                                          session_recording=kwargs.get('recording', None),
-                                          typescript_recording=kwargs.get('typescriptrecording', None))
-
-            # admin parameter is optional yet if not set connections may fail
-            admin_name = kwargs.get('admin')
-            adm_rec = RecordMixin.resolve_single_record(params, admin_name)
-            admin_uid = adm_rec.record_uid if adm_rec else None
-            if admin_uid and record_type in ("pamDatabase", "pamDirectory", "pamMachine"):
-                tdag.link_user_to_resource(admin_uid, record_uid, is_admin=True, belongs_to=True)
-                # tdag.link_user_to_config(admin_uid)  # is_iam_user=True
-
-            # Print out PAM Settings
-            if not kwargs.get("silent", False): tdag.print_tunneling_config(record_uid, record.get_typed_field('pamSettings'), config_uid)
-
-
-class PAMSplitCommand(Command):
-    pam_cmd_parser = argparse.ArgumentParser(prog='pam split')
-    pam_cmd_parser.add_argument('pam_machine_record', type=str, action='store',
-                                help='The record UID or title of the legacy PAM Machine '
-                                'record with built-in PAM User credentials.')
-    pam_cmd_parser.add_argument('--configuration', '-c', required=False, dest='pam_config', action='store',
-                                help='The PAM Configuration Name or UID - If the legacy record was configured '
-                                     'for rotation this command will try to autodetect PAM Configuration settings '
-                                     'otherwise you\'ll be prompted to provide the PAM Config.')
-    pam_cmd_parser.add_argument('--folder', '-f', required=False, dest='pam_user_folder', action='store',
-                                help='The folder where to store the new PAM User record - '
-                                     'folder names/paths are case sensitive!'
-                                     '(if skipped - PAM User will be created into the '
-                                     'same folder as PAM Machine)')
-
-    def get_parser(self):
-        return PAMSplitCommand.pam_cmd_parser
-
-    def execute(self, params, **kwargs):
-        def remove_field(record, field): # type: (vault.TypedRecord, vault.TypedField) -> bool
-            # Since TypedRecord.get_typed_field scans both fields[] and custom[]
-            # we need corresponding remove field lookup
-            fld = next((x for x in record.fields if field.type == x.type and
-                        (not field.label or
-                        (x.label and field.label.casefold() == x.label.casefold()))), None)
-            if fld is not None:
-                record.fields.remove(field)
-                return True
-
-            fld = next((x for x in record.custom if field.type == x.type and
-                        (not field.label or
-                        (x.label and field.label.casefold() == x.label.casefold()))), None)
-            if fld is not None:
-                record.custom.remove(field)
-                return True
-
-            return False
-
-        def resolve_record(params, name):
-            record_uid = None
-            if name in params.record_cache:
-                record_uid = name  # unique record UID
-            else:
-                # lookup unique folder/record path
-                rs = try_resolve_path(params, name)
-                if rs is not None:
-                    folder, name = rs
-                    if folder is not None and name is not None:
-                        folder_uid = folder.uid or ''
-                        if folder_uid in params.subfolder_record_cache:
-                            for uid in params.subfolder_record_cache[folder_uid]:
-                                r = api.get_record(params, uid)
-                                if r.title.lower() == name.lower():
-                                    record_uid = uid
-                                    break
-            if not record_uid:
-                # lookup unique record title
-                records = []
-                for uid in params.record_cache:
-                    data_json = params.record_cache[uid].get("data_unencrypted", "{}") or {}
-                    data = json.loads(data_json)
-                    if "pamMachine" == str(data.get("type", "")):
-                        title = data.get('title', '') or ''
-                        if title.lower() == name.lower():
-                            records.append(uid)
-                uniq_recs = len(set(records))
-                if uniq_recs > 1:
-                    print(f"{bcolors.FAIL}Multiple PAM Machine records match title '{name}' - "
-                          f"specify unique record path/name.{bcolors.ENDC}")
-                elif records:
-                    record_uid = records[0]
-            return record_uid
-
-        def resolve_folder(params, name):
-            folder_uid = ''
-            if name:
-                # lookup unique folder path
-                folder_uid = FolderMixin.resolve_folder(params, name)
-                # lookup unique folder name/uid
-                if not folder_uid and name != '/':
-                    folders = []
-                    for fkey in params.subfolder_cache:
-                        data_json = params.subfolder_cache[fkey].get('data_unencrypted', '{}') or {}
-                        data = json.loads(data_json)
-                        fname = data.get('name', '') or ''
-                        if fname == name:
-                            folders.append(fkey)
-                    uniq_items = len(set(folders))
-                    if uniq_items > 1:
-                        print(f"{bcolors.FAIL}Multiple folders match '{name}' - specify unique "
-                                f"folder name or use folder UID (or omit --folder parameter to create "
-                                f"PAM User record in same folder as PAM Machine record).{bcolors.ENDC}")
-                        folders = []
-                    folder_uid = folders[0] if folders else ''
-            return folder_uid
-
-        def resolve_pam_config(params, record_uid, pam_config_option):
-            # PAM Config lookup - Legacy PAM Machine will have associated PAM Config
-            # only if it is set up for rotation - otherwise PAM Config must be provided
-            encrypted_session_token, encrypted_transmission_key, transmission_key = get_keeper_tokens(params)
-            pamcfg_rec = get_config_uid(params, encrypted_session_token, encrypted_transmission_key, record_uid)
-            if not pamcfg_rec and not pam_config_option:
-                print(f"{bcolors.FAIL}Unable to find PAM Config associated with record '{record_uid}' "
-                    "- please provide PAM Config with --configuration|-c option. "
-                    "(Note: Legacy PAM Machine is linked to PAM Config only if "
-                    f"the machine is set up for rotation).{bcolors.ENDC}")
-                return
-
-            pamcfg_cmd = ''
-            if pam_config_option:
-                pam_uids = []
-                for uid in params.record_cache:
-                    if params.record_cache[uid].get('version', 0) == 6:
-                        r = api.get_record(params, uid)
-                        if r.record_uid == pam_config_option or r.title.lower() == pam_config_option.lower():
-                            pam_uids.append(uid)
-                uniq_recs = len(set(pam_uids))
-                if uniq_recs > 1:
-                    print(f"{bcolors.FAIL}Multiple PAM Config records match '{pam_config_option}' - "
-                            f"specify unique record UID/Title.{bcolors.ENDC}")
-                elif pam_uids:
-                    pamcfg_cmd = pam_uids[0]
-                elif not pamcfg_rec:
-                    print(f"{bcolors.FAIL}Unable to find PAM Configuration '{pam_config_option}'.{bcolors.ENDC}")
-
-            # PAM Config set on command line overrides the PAM Machine associated PAM Config
-            pam_config_uid = pamcfg_cmd or pamcfg_rec or ""
-            if pamcfg_cmd and pamcfg_rec and pamcfg_cmd != pamcfg_rec:
-                print(f"{bcolors.WARNING}PAM Config associated with record '{record_uid}' "
-                    "is different from PAM Config set with --configuration|-c option. "
-                    f"Using the configuration from command line option.{bcolors.ENDC}")
-
-            return pam_config_uid
-
-        # Parse command params
-        pam_config = kwargs.get('pam_config', '')  # PAM Configuration Name or UID
-        folder = kwargs.get('pam_user_folder', '')  # destination folder
-        record_uid = kwargs.get('pam_machine_record', '')  # existing record UID
-
-        record_uid = resolve_record(params, record_uid) or record_uid
-        record = vault.KeeperRecord.load(params, record_uid)
-        if not record:
-            raise CommandError('', f"{bcolors.FAIL}Record {record_uid} not found.{bcolors.ENDC}")
-        if not isinstance(record, vault.TypedRecord) or record.record_type != "pamMachine":
-            raise CommandError('', f"{bcolors.FAIL}Record {record_uid} is not of the expected type 'pamMachine'.{bcolors.ENDC}")
-
-        pam_config_uid = resolve_pam_config(params, record_uid, pam_config)
-        if not pam_config_uid:
-            print(f"{bcolors.FAIL}Please provide a valid PAM Configuration.{bcolors.ENDC}")
-            return
-            # print(f"{bcolors.WARNING}Failed to find PAM Configuration for {record_uid} "
-            #       "and unable to link new PAM User to PAM Machine. Remember to manually link "
-            #       f"Administrative Credentials record later.{bcolors.ENDC}")
-
-        folder_uid = resolve_folder(params, folder)
-        if folder and not folder_uid:
-            print(f"{bcolors.WARNING}Unable to find destination folder '{folder}' "
-                  "(Note: folder names/paths are case sensitive) "
-                  "- PAM User record will be stored into same folder "
-                  f"as the originating PAM Machine record.{bcolors.ENDC}")
-
-        flogin = record.get_typed_field('login')
-        vlogin = flogin.get_default_value(str) if flogin else ''
-        fpass = record.get_typed_field('password')
-        vpass = fpass.get_default_value(str) if fpass else ''
-        fpkey = record.get_typed_field('secret')
-        vpkey = fpkey.get_default_value(str) if fpkey else ''
-        if not(vlogin or vpass or vpkey):
-            if not(flogin or fpass or fpkey):
-                print(f"{bcolors.WARNING}Record {record_uid} is already in the new format.{bcolors.ENDC}")
-            else:
-                # No values present - just drop the old fields and add new ones
-                # thus converting the record to the new pamMachine format
-                # NB! If record was edited - newer clients moved these to custom fields
-                if flogin:
-                    remove_field(record, flogin)
-                if fpass:
-                    remove_field(record, fpass)
-                if fpkey:
-                    remove_field(record, fpkey)
-
-                if not record.get_typed_field('trafficEncryptionSeed'):
-                    record_seed = vault.TypedField.new_field('trafficEncryptionSeed', "", "")
-                    record.fields.append(record_seed)
-                if not record.get_typed_field('pamSettings'):
-                    pam_settings = vault.TypedField.new_field('pamSettings', "", "")
-                    record.fields.append(pam_settings)
-
-                record_management.update_record(params, record)
-                params.sync_data = True
-
-                print(f"{bcolors.WARNING}Record {record_uid} has no data to split and "
-                    "was converted to the new format. Remember to manually add "
-                    f"Administrative Credentials later.{bcolors.ENDC}")
-            return
-        elif not vlogin or not(vpass or vpkey):
-            print(f"{bcolors.WARNING}Record {record_uid} has incomplete user data "
-                  "but splitting anyway. Remember to manually update linked "
-                  f"Administrative Credentials record later.{bcolors.ENDC}")
-
-        # Create new pamUser record
-        user_rec = vault.KeeperRecord.create(params, 'pamUser')
-        user_rec.type_name = 'pamUser'
-        user_rec.title = str(record.title) + ' Admin User'
-        if flogin:
-            field = user_rec.get_typed_field('login')
-            field.value = flogin.value
-        if fpass:
-            field = user_rec.get_typed_field('password')
-            field.value = fpass.value
-        if fpkey:
-            field = user_rec.get_typed_field('secret')
-            field.value = fpkey.value
-
-        if not folder_uid:  # use the folder of the PAM Machine record
-            folders = list(find_folders(params, record.record_uid))
-            uniq_items = len(set(folders))
-            if uniq_items < 1:
-                print(f"{bcolors.WARNING}The new record will be created in root folder.{bcolors.ENDC}")
-            elif uniq_items > 1:
-                print(f"{bcolors.FAIL}Record '{record.record_uid}' is probably "
-                      "a linked record with copies/links across multiple folders "
-                      f"and PAM User record will be created in folder '{folders[0]}'.{bcolors.ENDC}")
-            folder_uid = folders[0] if folders else ''  # '' means root folder
-
-        record_management.add_record_to_folder(params, user_rec, folder_uid)
-        pam_user_uid = params.environment_variables.get(LAST_RECORD_UID, '')
-        api.sync_down(params)
-
-        if flogin:
-            remove_field(record, flogin)
-        if fpass:
-            remove_field(record, fpass)
-        if fpkey:
-            remove_field(record, fpkey)
-
-        if not record.get_typed_field('trafficEncryptionSeed'):
-            record_seed = vault.TypedField.new_field('trafficEncryptionSeed', "", "")
-            record.fields.append(record_seed)
-        if not record.get_typed_field('pamSettings'):
-            pam_settings = vault.TypedField.new_field('pamSettings', "", "")
-            record.fields.append(pam_settings)
-
-        record_management.update_record(params, record)
-        params.sync_data = True
-
-        if pam_config_uid:
-            encrypted_session_token, encrypted_transmission_key, transmission_key = get_keeper_tokens(params)
-            tdag = TunnelDAG(params, encrypted_session_token, encrypted_transmission_key, pam_config_uid, True)
-            tdag.link_resource_to_config(record_uid)
-            tdag.link_user_to_resource(pam_user_uid, record_uid, True, True)
-
-        print(f"PAM Machine record {record_uid} user credentials were split into "
-              f"a new PAM User record {pam_user_uid}")
-
-        return
