+++ keepercommander/params.py
@@ -161,8 +161,7 @@ class KeeperParams:
         self.ssh_agent = None
         self.unmask_all = False
         self.ws = None
-        self.tunnel_threads = {}
-        self.tunnel_threads_queue = {} # add ability to tail tunnel process
+        self.tube_registry = None  # Rust WebRTC tube registry instance
         self.forbid_rsa = False
         # TODO check if it can be deleted
         self.salt = None
@@ -230,8 +229,12 @@ class KeeperParams:
         if self.ssh_agent:
             self.ssh_agent.close()
             self.ssh_agent = None
-        self.tunnel_threads.clear()
-        self.tunnel_threads_queue = {}
+        if self.tube_registry:
+            try:
+                self.tube_registry.cleanup_all()
+            except Exception:
+                pass  # Ignore cleanup errors during session clear
+            self.tube_registry = None
         self.forbid_rsa = False
 
     def __get_rest_context(self):   # type: () -> RestApiContext
