+++ setup.cfg
@@ -39,11 +39,9 @@ install_requires =
     pycryptodomex>=3.20.0
     pyngrok; python_version>='3.8'
     pyperclip
-    pysocks
     python-dotenv
     requests>=2.31.0
     tabulate
-    websockets
     aiortc; python_version>='3.8'
     pydantic>=2.6.4; python_version>='3.8'
     fpdf2>=2.8.3
